+++ src/routing/or.rs
@@ -10,7 +10,8 @@ use std::{
     pin::Pin,
     task::{Context, Poll},
 };
-use tower::{util::Oneshot, Service, ServiceExt};
+use tower::{util::Oneshot, ServiceExt};
+use tower_service::Service;
 
 /// [`tower::Service`] that is the combination of two routers.
 ///